gui/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
audio-recorder-streamlit>=0.0.8
